Move token list and action handling to new AppModel
